+++ components/game/game.test.tsx
@@ -3,9 +3,7 @@ import '@testing-library/jest-dom';
 import GameLayout from './game';
 import { mockGameQuestion } from '../../mock/game.mock';
 import { Provider } from 'react-redux';
-import { configureStore } from '@reduxjs/toolkit';
-import gameSliceMock from '../../store/game-slice.mock';
-import { gameMockEmptySlice } from '../../store/game-slice.mock';
+import { createTestStore } from '../../store/fake-store';
 
 jest.mock("next-auth/react", () => {
   const originalModule = jest.requireActual('next-auth/react');
@@ -27,26 +25,6 @@ jest.mock('next/router', () => ({
     playlistId: 'test'
   }}),
 }));
-
-const createTestStore = () => {
-  const store = configureStore({
-    reducer: {
-      game: gameSliceMock
-    }
-  });
-  return store;
-};
-
-const createEmptyTestStore = () => {
-  const store = configureStore({
-    reducer: {
-      game: gameMockEmptySlice.reducer
-    }
-  });
-  return store;
-};
-
-
 
 describe('GameLayout', () => {
   let fakeStore: any;
@@ -79,25 +57,4 @@ describe('GameLayout', () => {
     expect(endGameElement).toBeInTheDocument();
     expect(nextElement).toBeInTheDocument();
   });
-});
-
-describe('Empty state game', () => {
-  let fakeStore: any;
-  beforeEach(() => {
-    fakeStore = createEmptyTestStore();
-  })
-
-  beforeEach(() => {
-    render(
-      <Provider store={fakeStore}>
-        <GameLayout />;
-      </Provider>
-    );
-  });
-
-  it('Should render loading message', () => {
-    const loadingMessage = screen.getByText('Loading ...');
-    expect(loadingMessage).toBeInTheDocument();
-  })
-
 });
