+++ components/game/playlist.test.tsx
@@ -1,15 +1,25 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
+import { Provider } from 'react-redux';
 import PlaylistLayout from './playlists';
 import { playlistsMock } from '../../mock/playlist.mock';
+import { createTestStore } from '../../store/fake-store';
 
 jest.mock('next/router', () => ({
   useRouter: jest.fn(),
 }));
 
 describe('PlaylistLayout', () => {
+  let fakeStore: any;
   beforeEach(() => {
-    render(<PlaylistLayout playlists={playlistsMock} />);
+    fakeStore = createTestStore();
+  })
+  beforeEach(() => {
+    render(
+      <Provider store={fakeStore}>
+        <PlaylistLayout playlists={playlistsMock} />
+      </Provider>
+    );
   });
 
   it('Should render all the playlists', () => {
