+++ components/menu/menu.test.tsx
@@ -1,6 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
+import { Provider } from 'react-redux';
 import MenuLayout from './menu';
+import { createTestStore } from '../../store/fake-store';
 
 jest.mock('next/router', () => ({
   useRouter: jest.fn(),
@@ -10,8 +12,17 @@ const actions = ['Play'];
 
 describe('Menu', () => {
 
+  let fakeStore: any;
   beforeEach(() => {
-    render(<MenuLayout/>);
+    fakeStore = createTestStore();
+  })
+
+  beforeEach(() => {
+    render(
+      <Provider store={fakeStore}>
+        <MenuLayout/>
+      </Provider>
+    );
   })
 
   it('Should render logos', () => {
